+++ config-updater/scripts/functions.sh
@@ -59,6 +59,7 @@ _update_config() {
     sed -i 's/\r$//g' -- *.ovpn
     sed -i 's/^auth-user-pass$/auth-user-pass \/control\/ovpn-auth.txt/' -- *.ovpn
     sed -i -rE 's/^remote .+ ([[:digit:]]+)$/remote {{ .Env.OPENVPN_HOSTNAME }} \1/' -- *.ovpn
+    sed -i -rE 's/^verify-x509-name .+ name$/verify-x509-name {{ .Env.OPENVPN_HOSTNAME }} name/' -- *.ovpn
     if [[ ! -d "${parent_config_path}/${vpn_provider}/${proto}" ]]; then
         mkdir -p "${parent_config_path}/${vpn_provider}/${proto}"
     fi
@@ -76,6 +77,24 @@ _update_config() {
             sed -i 's/ping 15/inactive 3600\nping 10/g' default.ovpn.tmpl
             sed -i 's/ping-restart 0/ping-exit 60/g' default.ovpn.tmpl
             sed -i 's/ping-timer-rem//g' default.ovpn.tmpl
+        fi
+        ca_conf=$(grep -E '^ca .*' default.ovpn.tmpl)
+        if [[ -n "${ca_conf}" ]]; then
+            ca_file=$(echo "${ca_conf}" | awk '{ print $2 }')
+            mv "${ca_file}" "${parent_config_path}/${vpn_provider}/${proto}/${ca_file}"
+            sed -i -rE "s/^ca .*?$/ca \/etc\/templates\/openvpn\/${vpn_provider}\/${proto}\/${ca_file}/" -- default.ovpn.tmpl
+        fi
+        cert_conf=$(grep -E '^cert .*' default.ovpn.tmpl)
+        if [[ -n "${cert_conf}" ]]; then
+            cert_file=$(echo "${cert_conf}" | awk '{ print $2 }')
+            mv "${cert_file}" "${parent_config_path}/${vpn_provider}/${proto}/${cert_file}"
+            sed -i -rE "s/^cert .*?$/cert \/etc\/templates\/openvpn\/${vpn_provider}\/${proto}\/${cert_file}/" -- default.ovpn.tmpl
+        fi
+        key_conf=$(grep -E '^key .*' default.ovpn.tmpl)
+        if [[ -n "${key_conf}" ]]; then
+            key_file=$(echo "${key_conf}" | awk '{ print $2 }')
+            mv "${key_file}" "${parent_config_path}/${vpn_provider}/${proto}/${key_file}"
+            sed -i -rE "s/^key .*?$/key \/etc\/templates\/openvpn\/${vpn_provider}\/${proto}\/${key_file}/" -- default.ovpn.tmpl
         fi
         mv default.ovpn.tmpl "${parent_config_path}/${vpn_provider}/${proto}/default.ovpn.tmpl"
     fi
@@ -217,4 +236,77 @@ _update_pia_config() {
         fi
     done
     popd > /dev/null || exit
+}
+
+# Get VyprVPN configs and update
+_update_vyprvpn_config() {
+    local vpn_provider=$1
+    local config_url=$2
+    local extra_pattern=$3
+    local tmp_dir target_dir
+    echo "Getting VyprVPN configs using ${config_url}..."
+    tmp_dir=$(mktemp -d /tmp/vpn.XXXXXXXX)
+    target_dir=$(mktemp -d /tmp/target.XXXXXXXX)
+    mkdir "${target_dir}"/{tcp,udp}
+    pushd "${tmp_dir}" > /dev/null || exit
+    curl -4 -sSL "${config_url}" -o vyprvpn.zip || exit
+    unzip -q vyprvpn.zip || exit
+    find . -name "*.ovpn" -regex ".*${extra_pattern}.*" -exec bash -c 'mv "${1}" "${0}/$(basename ${1// /_})"' "${target_dir}/udp" {} \;
+    popd > /dev/null || exit
+    _pre_config_update "${vpn_provider}" "${target_dir}" "udp"
+}
+
+# Get SurfShark configs and update
+_update_surfshark_config() {
+    local vpn_provider=$1
+    local config_url=$2
+    local tmp_dir target_dir
+    echo "Getting SurfShark configs using ${config_url}..."
+    tmp_dir=$(mktemp -d /tmp/vpn.XXXXXXXX)
+    target_dir=$(mktemp -d /tmp/target.XXXXXXXX)
+    mkdir "${target_dir}"/{tcp,udp}
+    pushd "${tmp_dir}" > /dev/null || exit
+    curl -4 -sSL "${config_url}" -o surfshark.zip || exit
+    unzip -q surfshark.zip || exit
+    find . -name "*udp*.ovpn" -exec mv {} "${target_dir}/udp/" \;
+    find . -name "*tcp*.ovpn" -exec mv {} "${target_dir}/tcp/" \;
+    popd > /dev/null || exit
+    _pre_config_update "${vpn_provider}" "${target_dir}"
+}
+
+# Get IPVanish configs and update
+_update_ipvanish_config() {
+    local vpn_provider=$1
+    local config_url=$2
+    local tmp_dir target_dir
+    echo "Getting IPVanish configs using ${config_url}..."
+    tmp_dir=$(mktemp -d /tmp/vpn.XXXXXXXX)
+    target_dir=$(mktemp -d /tmp/target.XXXXXXXX)
+    mkdir "${target_dir}"/{tcp,udp}
+    pushd "${tmp_dir}" > /dev/null || exit
+    curl -4 -sSL "${config_url}" -o ipvanish.zip || exit
+    unzip -q ipvanish.zip || exit
+    find . \( -name "*.ovpn" -o -name "*.crt" \) -exec mv {} "${target_dir}/udp/" \;
+    popd > /dev/null || exit
+    _pre_config_update "${vpn_provider}" "${target_dir}" "udp"
+}
+
+# Get TunnelBear configs and update
+_update_tunnelbear_config() {
+    local vpn_provider=$1
+    local config_url=$2
+    local tmp_dir target_dir
+    echo "Getting TunnelBear configs using ${config_url}..."
+    tmp_dir=$(mktemp -d /tmp/vpn.XXXXXXXX)
+    target_dir=$(mktemp -d /tmp/target.XXXXXXXX)
+    mkdir "${target_dir}"/{tcp,udp}
+    pushd "${tmp_dir}" > /dev/null || exit
+    curl -4 -sSL "${config_url}" -o tunnelbear.zip || exit
+    unzip -q tunnelbear.zip || exit
+    if [[ -d "openvpn" ]]; then
+        cd openvpn || exit
+    fi
+    find . \( -name "*.ovpn" -o -name "*.crt" -o -name "*.key" \) -exec bash -c 'mv "${1}" "${0}/$(basename ${1// /_})"' "${target_dir}/udp" {} \;
+    popd > /dev/null || exit
+    _pre_config_update "${vpn_provider}" "${target_dir}" "udp"
 }