+++ config-updater/scripts/update-config.sh
@@ -38,6 +38,14 @@ elif [[ "${VPN_PROVIDER,,}" == "purevpn" ]]; then
     _update_purevpn_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}"
 elif [[ "${VPN_PROVIDER,,}" == "pia" ]]; then
     _update_pia_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}"
+elif [[ "${VPN_PROVIDER,,}" == "vyprvpn" ]]; then
+    _update_vyprvpn_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}" "256"
+elif [[ "${VPN_PROVIDER,,}" == "surfshark" ]]; then
+    _update_surfshark_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}"
+elif [[ "${VPN_PROVIDER,,}" == "ipvanish" ]]; then
+    _update_ipvanish_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}"
+elif [[ "${VPN_PROVIDER,,}" == "tunnelbear" ]]; then
+    _update_tunnelbear_config "${VPN_PROVIDER,,}" "${PROVIDER_URL}"
 fi
 
 rm -rf "${TMP_DIR}"